Revise more documentation for UTSound
